% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flag.R
\name{draw_key_flag}
\alias{draw_key_flag}
\title{Draw flag key}
\usage{
draw_key_flag(country = "nz")
}
\arguments{
\item{country}{A country whose fla gwill be used for the legend key}
}
\value{
A function that draws the flag key
}
\description{
Draw flag key
}
