% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flag.R
\name{geom_flag}
\alias{geom_flag}
\title{geom_flag}
\usage{
geom_flag(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{stat}{}

\item{position}{}

\item{na.rm}{}

\item{show.legend}{}

\item{inherit.aes}{}

\item{...}{}
}
\description{
geom_flag
}
\examples{
data(lflags)
set.seed(1234)
d <- data.frame(x=rnorm(10), y=rnorm(10), 
                country=sample(c("ar","fr"), 10, TRUE), 
                stringsAsFactors = FALSE)
ggplot(d, aes(x=x, y=y, country=country, size=x)) + 
  geom_flag() + 
  scale_country()
}

