% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flag_interactive.R
\name{geom_flag_interactive}
\alias{geom_flag_interactive}
\title{Create interactive flags}
\usage{
geom_flag_interactive(...)
}
\arguments{
\item{...}{arguments passed to base function,
plus any of the [ggiraph:::interactive_parameters()].}
}
\description{
An interactive version of [geom_flag()], designed to work with the `ggiraph`
package.
}
\examples{
# add interactive flags to a ggplot -------
}
\seealso{
[ggiraph::girafe()]
}
