% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flag_interactive.R
\name{interactive_flags_grob}
\alias{interactive_flags_grob}
\title{Create interactive flags grob}
\usage{
interactive_flags_grob(...)
}
\arguments{
\item{...}{arguments passed to base function,
plus any of the [interactive_parameters()].}
}
\value{
An interactive grob object.
}
\description{
The grob is based on [flagsGrob()].
See the documentation for that function for more details.
}
\seealso{
[girafe()]
}
